// src/app/page/master/map-kelompok-user-jabatan/map-kelompok-user-jabatan.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapKelompokUserJabatan {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	values?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}